refactor: update WageWithForm import paths for consistency across components
